id'ification of vertices and separation of graph from history.
Generating 2D layouts for graphs was very time consuming.  Now you can
do that independently of generating the history for your search.

Dealing with indices was cumbersome and fragile.  Now vertices are
referred to by an id (any object with an unique string representation in
your graph).

Incidentally, this new setup now allows for displaying the same search
in different graphs.